server/requirements.txt
Flask
pymongo
flask-cors
python-dotenv
pytest
pytest-flask